More typefixing
